+++ packages/ui-react/src/lib/components/IconButton.tsx
@@ -21,7 +21,7 @@ export const IconButton = ({
   disabled = false,
   type = 'button',
   title,
-  ariaLabel,
+  label,
   onToggle,
   activated = false,
   onClick,
@@ -34,10 +34,17 @@ export const IconButton = ({
   shape = 'rounded',
   allowShapeTransformation = true,
   transition,
+  children,
   ...restProps
 }: ReactProps<IconButtonInterface>) => {
+  if (children) label = children;
+  if (!label) {
+    throw new Error(
+      'IconButton component requires either a label prop or children content to provide an accessible aria-label',
+    );
+  }
   if (!title) {
-    title = ariaLabel;
+    title = label;
   }
 
   const [isActive, setIsActive] = React.useState(activated);
@@ -76,7 +83,7 @@ export const IconButton = ({
     width,
     href,
     activated: isActive,
-    ariaLabel,
+    label,
     iconSelected,
     isActive,
     onToggle,
@@ -85,6 +92,7 @@ export const IconButton = ({
     variant,
     className,
     size,
+    children: label,
     ...restProps,
   });
 
@@ -97,8 +105,9 @@ export const IconButton = ({
     <ElementType
       disabled={disabled}
       href={href}
+      style={{ transition: transition.duration + 's' }}
       className={styles.iconButton}
-      aria-label={ariaLabel}
+      aria-label={label}
       {...(restProps as any)}
       title={undefined}
       onClick={handleClick}
@@ -109,38 +118,38 @@ export const IconButton = ({
         trigger={disabled ? null : undefined}
         text={title}
       ></ToolTip>
-      <div
-        style={{ transition: transition.duration + 's' }}
-        className={styles.container}
-      >
-        <div className={styles.stateLayer}>
-          {!disabled && (
-            <RippleEffect
-              colorName={classNames(
-                variant === 'standard' && {
-                  'on-surface-variant': !isActive,
-                  primary: isActive,
-                },
-                variant === 'filled' && {
-                  primary: !isActive && Boolean(onToggle),
-                  'inverse-on-surface': isActive || !onToggle,
-                },
-                variant === 'tonal' && {
-                  'on-surface-variant': !isActive && Boolean(onToggle),
-                  'on-secondary-container': isActive || !onToggle,
-                },
-                variant === 'outlined' && {
-                  'on-surface-variant': !isActive,
-                  'on-primary': isActive,
-                },
-              )}
-              triggerRef={resolvedRef}
-            />
-          )}
-        </div>
 
-        {icon && <Icon icon={icon} className={styles.icon} />}
+      <div className={styles.touchTarget} />
+      <div
+        className={styles.stateLayer}
+        style={{ transition: transition.duration + 's' }}
+      >
+        {!disabled && (
+          <RippleEffect
+            colorName={classNames(
+              variant === 'standard' && {
+                'on-surface-variant': !isActive,
+                primary: isActive,
+              },
+              variant === 'filled' && {
+                primary: !isActive && Boolean(onToggle),
+                'inverse-on-surface': isActive || !onToggle,
+              },
+              variant === 'tonal' && {
+                'on-surface-variant': !isActive && Boolean(onToggle),
+                'on-secondary-container': isActive || !onToggle,
+              },
+              variant === 'outlined' && {
+                'on-surface-variant': !isActive,
+                'on-primary': isActive,
+              },
+            )}
+            triggerRef={resolvedRef}
+          />
+        )}
       </div>
+
+      {icon && <Icon icon={icon} className={styles.icon} />}
     </ElementType>
   );
 };