+++ packages/ui-react/src/lib/interfaces/card.interface.ts
@@ -1,8 +1,11 @@
+import { ReactNode } from 'react';
+
 export interface CardInterface {
   type: 'div';
   props: {
     variant?: 'outlined' | 'elevated' | 'filled';
     isInteractive?: boolean;
+    children: ReactNode;
   };
   elements: ['card', 'stateLayer'];
 }