+++ packages/ui-react/src/lib/interfaces/slider.interface.ts
@@ -2,11 +2,9 @@ export interface SliderInterface {
   type: 'div';
   props: {
     value?: number;
-
     name: string;
     onChange?: (value: number) => void;
     valueFormatter?: (value: number) => string | number;
-
     step?: number;
     min?: number;
     max?: number;